// frontend/src/sanity/types.ts
export type CartItem = {
  _key: string;
  quantity: number;
  dish: {
    _id: string;
    name?: string;
    description?: string;
    price?: number;
    image?: SanityImage;
    restaurant?: {
      _id: string;
      name?: string;
      logo?: SanityImage;
    };
  };
};

export type CartOrder = {
  _id: string;
  items: CartItem[];
  totalPrice: number;
  status: string;
};

export type SanityImage = {
  _type: "image";
  asset: {
    _ref: string;
    _type: "reference";
  };
  crop?: {
    top: number;
    bottom: number;
    left: number;
    right: number;
  };
  hotspot?: {
    x: number;
    y: number;
    height: number;
    width: number;
  };
};